+++ tests/Unit/Events/PaymentReversedTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use CGrate\Laravel\Events\PaymentReversed;
+use CGrate\Laravel\Events\CashDeposit;
 use CGrate\Php\DTOs\ReversePaymentResponseDTO;
 use CGrate\Php\Enums\ResponseCode;
 use Illuminate\Foundation\Events\Dispatchable;
@@ -10,7 +10,7 @@ use Illuminate\Queue\SerializesModels;
 use Illuminate\Support\Facades\Event;
 
 it('uses Dispatchable and SerializesModels traits', function () {
-    $traits = class_uses(PaymentReversed::class);
+    $traits = class_uses(CashDeposit::class);
 
     expect($traits)->toHaveKey(Dispatchable::class);
     expect($traits)->toHaveKey(SerializesModels::class);
@@ -24,7 +24,7 @@ it('can be instantiated with reverse payment response and reference', function (
     ];
 
     $response = ReversePaymentResponseDTO::fromResponse($responseData);
-    $event = new PaymentReversed($response);
+    $event = new CashDeposit($response);
 
     expect($event->response)->toBe($response);
     expect($event->response->responseCode)->toBe(ResponseCode::SUCCESS);
@@ -40,9 +40,9 @@ it('can be dispatched and listened to', function () {
 
     $response = ReversePaymentResponseDTO::fromResponse($responseData);
 
-    Event::fake([PaymentReversed::class]);
-    PaymentReversed::dispatch($response);
-    Event::assertDispatched(PaymentReversed::class, function ($event) use ($response) {
+    Event::fake([CashDeposit::class]);
+    CashDeposit::dispatch($response);
+    Event::assertDispatched(CashDeposit::class, function ($event) use ($response) {
         return $event->response === $response;
     });
 });